+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Added space to include custom markup at the end of the component.
+
 ## [2.1.0] - 2022-04-12
 
 ### Added